Refactor classes, fix unit test cases.
Add validate command to cli

Rename and refactor.

Fix submit workflows

Only change the auth part of the submit
workflow, instead of also changing the
functionality. So that we keep the submit
function work.

Rename and refactor files

Update gitignore file to ignore testing cache.

Polish helper messages in the CLI.

Polish docstrings and add todo items for all of the classes.

Polish test cases.
