+++ cromwell_tools/cromwell_api.py
@@ -3,14 +3,16 @@ TODO: add some module docs
 """
 
 import time
-
+import tempfile
+from subprocess import PIPE, Popen
+import os
 import json
 import logging
 import requests
 from datetime import datetime, timedelta
 from tenacity import retry, stop_after_delay, wait_exponential
 
-from cromwell_tools._utilities import prepare_workflow_manifest, validate_cromwell_label
+from cromwell_tools.utilities import validate_cromwell_label, _localize_file
 
 
 logger = logging.getLogger(__name__)
@@ -26,26 +28,16 @@ class WorkflowUnknownException(Exception):
 
 # TODO: use functools partial for get, post (set the authenticate commands)
 class CromwellAPI(object):
-    """Contains a set of classmethods that implement interfaces to cromwell REST API endpoints.
-    """
-
-    # TODO: remove this if it is not critical for mocking
-    requests = requests
+    """Contains a set of classmethods that implement interfaces to cromwell REST API endpoints."""
 
     # TODO: move the endpoints definitions to the corresponding functions after refactoring the unit tests and mocks
     _abort_endpoint = '/api/workflows/v1/{uuid}/abort'
-
     _status_endpoint = '/api/workflows/v1/{uuid}/status'
-
-    _workflow_endpoint = '/api/workflows/v1'
-
+    _submit_endpoint = '/api/workflows/v1'
     _metadata_endpoint = '/api/workflows/v1/{uuid}/metadata'
-
     _health_endpoint = '/api/engine/v1/status'
-
-    _release_hold_endpoint = 'api/workflows/v1/{uuid}/releaseHold'
-
-    _query_endpoint = 'api/workflows/v1/query'
+    _release_hold_endpoint = '/api/workflows/v1/{uuid}/releaseHold'
+    _query_endpoint = '/api/workflows/v1/query'
 
     _failed_statuses = ('Failed', 'Aborted', 'Aborting')
 
@@ -62,7 +54,7 @@ class CromwellAPI(object):
 
         Args:
             uuid (str): A Cromwell workflow UUID, which is the workflow identifier.
-            auth (cromwell_tools._cromwell_auth.CromwellAuth): The authentication class holding headers or auth
+            auth (cromwell_tools.cromwell_auth.CromwellAuth): The authentication class holding headers or auth
                 information to a Cromwell server.
             raise_for_status (Optional[bool]): Whether to check and raise for status based on the response. (default
                 False)
@@ -74,9 +66,9 @@ class CromwellAPI(object):
         Returns:
             requests.Response: HTTP response from Cromwell.
         """
-        response = cls.requests.post(url=auth.url + cls._abort_endpoint.format(uuid=uuid),
-                                     auth=auth.auth,
-                                     headers=auth.header)
+        response = requests.post(url=auth.url + cls._abort_endpoint.format(uuid=uuid),
+                                 auth=auth.auth,
+                                 headers=auth.header)
 
         if raise_for_status:
             cls._check_and_raise_status(response)
@@ -88,7 +80,7 @@ class CromwellAPI(object):
 
         Args:
             uuid (str): A Cromwell workflow UUID, which is the workflow identifier.
-            auth (cromwell_tools._cromwell_auth.CromwellAuth): The authentication class holding headers or auth
+            auth (cromwell_tools.cromwell_auth.CromwellAuth): The authentication class holding headers or auth
                 information to a Cromwell server.
             raise_for_status (Optional[bool]): Whether to check and raise for status based on the response. (default
                 False)
@@ -100,9 +92,9 @@ class CromwellAPI(object):
         Returns:
             requests.Response: HTTP response from Cromwell.
         """
-        response = cls.requests.get(url=auth.url + cls._status_endpoint.format(uuid=uuid),
-                                    auth=auth.auth,
-                                    headers=auth.header)
+        response = requests.get(url=auth.url + cls._status_endpoint.format(uuid=uuid),
+                                auth=auth.auth,
+                                headers=auth.header)
 
         if raise_for_status:
             cls._check_and_raise_status(response)
@@ -113,7 +105,7 @@ class CromwellAPI(object):
         """Return the current health status of any monitored subsystems of the Cromwell Server.
 
         Args:
-            auth (cromwell_tools._cromwell_auth.CromwellAuth): authentication class holding headers or auth
+            auth (cromwell_tools.cromwell_auth.CromwellAuth): authentication class holding headers or auth
                 information to a Cromwell server.
             raise_for_status (Optional[bool]): Whether to check and raise for status based on the response. (default
                 False)
@@ -125,9 +117,9 @@ class CromwellAPI(object):
         Returns:
             requests.Response: HTTP response from Cromwell.
         """
-        response = cls.requests.get(url=auth.url + cls._health_endpoint,
-                                    auth=auth.auth,
-                                    headers=auth.header)
+        response = requests.get(url=auth.url + cls._health_endpoint,
+                                auth=auth.auth,
+                                headers=auth.header)
 
         if raise_for_status:
             cls._check_and_raise_status(response)
@@ -135,27 +127,27 @@ class CromwellAPI(object):
 
     @classmethod
     @retry(reraise=True, wait=wait_exponential(multiplier=1, max=10), stop=stop_after_delay(20))
-    def submit(cls, auth, wdl_file, inputs_json, options_json=None, inputs2_json=None, dependencies_json=None,
-               collection_name=None, label=None, validate_labels=False):
-        """Submits a workflow to Cromwell.
-
-        # TODO: Allow more inputs files rather 2 to be consistent with the Cromwell API
-        # TODO: Add `raise_for_status` flag to this method and remove the hard-coded `raise_for_status()` call
-        # TODO: Purifying the methods by taking the specific retry policy out from this method
+    def submit(cls, auth, wdl_file, inputs_file, options_file=None, inputs_file2=None, zip_file=None,
+               collection_name=None, label=None, validate_labels=False, on_hold=False):
+        # TODO: Allow more inputs files rather 2 to be consistent with the Cromwell API.
+        # TODO: Add `raise_for_status` flag to this method and remove the hard-coded `raise_for_status()` call.
+        # TODO: Purifying the methods by taking the specific retry policy out from this method.
+        # TODO: Add back `prepare_workflow_manifest` to simplify this method and allow file paths as inputs.
+        """ Submits a workflow to Cromwell.
 
         Args:
-            wdl_file (_io.BytesIO or str): The workflow source file to submit for execution. From version 35,
+            auth (cromwell_tools.cromwell_auth.CromwellAuth): authentication class holding auth information to a Cromwell server.
+            wdl_file (_io.BytesIO): The workflow source file to submit for execution. From version 35,
                 Cromwell starts to accept URL to the WDL file besides actual WDL files.
-            inputs_json (_io.BytesIO): File-like object containing input data in JSON format.
-            options_json (Optional[_io.BytesIO]): Cromwell configs file.
-            inputs2_json (Optional[_io.BytesIO]): Inputs file 2.
-            dependencies_json (Optional[_io.BytesIO]): Zip file containing dependencies.
-            auth (cromwell_tools._cromwell_auth.CromwellAuth): authentication class holding headers or
-                auth information to a Cromwell server.
+            inputs_file (_io.BytesIO): File-like object containing input data in JSON format.
+            options_file (Optional[_io.BytesIO]): Cromwell configs file.
+            inputs_file2 (Optional[_io.BytesIO]): Inputs file 2.
+            zip_file (Optional[_io.BytesIO]): Zip file containing dependencies.
             collection_name (Optional[str]): Collection in SAM that the workflow should belong to.
             label (Optional[Union[str, _io.BytesIO]]): JSON file containing a collection of
                 key/value pairs for workflow labels. # TODO: verify these types are accurate
             validate_labels (Optional[bool]) If True, validate cromwell labels. (default False)
+            on_hold (Optional[bool]) Whether to submit the workflow in "On Hold" status.
 
         Raises:
             requests.exceptions.HTTPError: This will be raised when raise_for_status is True and Cromwell returns
@@ -164,35 +156,27 @@ class CromwellAPI(object):
         Returns:
             requests.Response: HTTP response from Cromwell.
         """
-        manifest = prepare_workflow_manifest(
-                wdl_file, inputs_json, dependencies_json, options_json, inputs2_json)
-
         if validate_labels and label is not None:
             validate_cromwell_label(label)
 
         files = {
             'workflowSource': wdl_file,
-            'workflowInputs': inputs_json,
+            'workflowInputs': inputs_file,
+            'workflowOnHold': json.dumps(on_hold)
         }
 
-        # TODO: these three checks should be done inside prepare workflow manifest?
-        if 'inputs2_json' in manifest:
-            files['workflowInputs_2'] = manifest['inputs2_json']
-        if 'dependencies_json' in manifest:
-            files['workflowDependencies'] = manifest['dependencies_json']
-        if 'options_json' in manifest:
-            files['workflowOptions'] = manifest['options_json']
-
+        if inputs_file2 is not None:
+            files['workflowInputs_2'] = inputs_file2
+        if zip_file is not None:
+            files['workflowDependencies'] = zip_file
+        if options_file is not None:
+            files['workflowOptions'] = options_file
         if label:
             files['labels'] = label
-        if collection_name:  # TODO: this check should happen elsewhere, related to caas
+        if collection_name:
             files['collectionName'] = collection_name
 
-        response = cls.requests.post(url=auth.url + cls._workflow_endpoint,
-                                     files=files,
-                                     auth=auth.auth,
-                                     headers=auth.header)
-
+        response = requests.post(auth.url + cls._submit_endpoint, files=files, auth=auth.auth, headers=auth.header)
         response.raise_for_status()
         return response
 
@@ -249,7 +233,7 @@ class CromwellAPI(object):
         Args:
             uuid: A Cromwell workflow UUID, which is the workflow identifier. The workflow is expected to have
                 `On Hold` status.
-            auth (cromwell_tools._cromwell_auth.CromwellAuth): The authentication class holding headers or auth
+            auth (cromwell_tools.cromwell_auth.CromwellAuth): The authentication class holding headers or auth
                 information to a Cromwell server.
             raise_for_status (Optional[bool]): Whether to check and raise for status based on the response. (default
                 False)
@@ -261,9 +245,9 @@ class CromwellAPI(object):
         Returns:
             response (requests.Response): HTTP response from Cromwell.
         """
-        response = cls.requests.post(url=auth.url + cls._release_hold_endpoint.format(uuid=uuid),
-                                     auth=auth.auth,
-                                     headers=auth.header)
+        response = requests.post(url=auth.url + cls._release_hold_endpoint.format(uuid=uuid),
+                                 auth=auth.auth,
+                                 headers=auth.header)
         if raise_for_status:
             cls._check_and_raise_status(response)
         return response
@@ -283,7 +267,7 @@ class CromwellAPI(object):
         Args:
             query_dict (dict): A dictionary representing the query key-value paris. The keys should be accepted by the
                 Cromwell or they will get ignored. The values could be str, list or dict.
-            auth (cromwell_tools._cromwell_auth.CromwellAuth): The authentication class holding headers or auth
+            auth (cromwell_tools.cromwell_auth.CromwellAuth): The authentication class holding headers or auth
                 information to a Cromwell server.
             raise_for_status (Optional[bool]): Whether to check and raise for status based on the response. (default
                 False)
@@ -296,10 +280,10 @@ class CromwellAPI(object):
             response (requests.Response): HTTP response from Cromwell.
         """
         query_params = cls._compose_query_params(query_dict)
-        response = cls.requests.post(url=auth.url + cls._query_endpoint,
-                                     json=query_params,
-                                     auth=auth.auth,
-                                     headers=auth.header)
+        response = requests.post(url=auth.url + cls._query_endpoint,
+                                 json=query_params,
+                                 auth=auth.auth,
+                                 headers=auth.header)
         if raise_for_status:
             cls._check_and_raise_status(response)
         return response
@@ -424,3 +408,32 @@ class CromwellAPI(object):
         """
         if not response.ok:
             raise requests.exceptions.HTTPError('Error Code {0}: {1}'.format(response.status_code, response.text))
+
+    @classmethod
+    def validate_workflow(cls, wdl, womtool_path, dependencies_json=None):
+        """Validate a wdl workflow using cromwell womtool.
+
+        Args:
+            wdl (str): Link or file path to wdl file.
+            womtool_path (str): Path to the womtool.jar. For more details about what is a womtools,
+                see https://github.com/broadinstitute/cromwell/releases)
+            dependencies_json (str): File path to json file containing workflow dependencies.
+        """
+        temporary_directory = tempfile.mkdtemp()
+
+        _localize_file(wdl, temporary_directory)
+        wdl_basename = os.path.basename(wdl)
+
+        if dependencies_json is not None:
+            with open(dependencies_json, 'r') as f:
+                depenencies_map = json.load(f)
+            for url in depenencies_map.values():
+                _localize_file(url, temporary_directory)
+
+        os.chdir(temporary_directory)
+        p = Popen(['java', '-jar', os.path.expanduser(womtool_path), 'validate', wdl_basename],
+                  stdout=PIPE, stderr=PIPE)
+        out, err = p.communicate()
+        if err:
+            raise ChildProcessError(err)
+        print('stdout:\n%s' % out.decode())
